old/python/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
docker==6.0.1
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
websocket-client==1.4.2
Werkzeug==2.2.2